refactor: move disableCssModuleExtension config to uni-builder
