+++ .changeset/strange-paws-yell.md
@@ -1,0 +1,8 @@
+---
+'@rsbuild/uni-builder': patch
+'@rsbuild/document': patch
+'@rsbuild/shared': patch
+'@rsbuild/core': patch
+---
+
+refactor: move disableCssModuleExtension config to uni-builder
