+++ packages/document/docs/en/config/options/output.mdx
@@ -48,12 +48,6 @@ import DisableCssExtract from '@en/shared/config/output/disableCssExtract.md';
 
 <DisableCssExtract />
 
-## output.disableCssModuleExtension
-
-import DisableCssModuleExtension from '@en/shared/config/output/disableCssModuleExtension.md';
-
-<DisableCssModuleExtension />
-
 ## output.distPath
 
 import DistPath from '@en/shared/config/output/distPath.md';
