+++ packages/document/docs/zh/config/options/output.mdx
@@ -48,12 +48,6 @@ import DisableCssExtract from '@zh/shared/config/output/disableCssExtract.md';
 
 <DisableCssExtract />
 
-## output.disableCssModuleExtension
-
-import DisableCssModuleExtension from '@zh/shared/config/output/disableCssModuleExtension.md';
-
-<DisableCssModuleExtension />
-
 ## output.distPath
 
 import DistPath from '@zh/shared/config/output/distPath.md';
