+++ packages/document/docs/zh/guide/basic/css-modules.md
@@ -34,36 +34,13 @@ export default () => {
 };
 ```
 
-## 为所有样式文件启用 CSS Modules
+## CSS Modules 识别规则
 
 在默认情况下，只有 `*.module.css` 结尾的文件才被视为 CSS Modules 模块。
 
-如果你想将源码目录下的所有 CSS 文件当做 CSS Modules 模块进行处理，可以通过开启 [output.disableCssModuleExtension](https://rsbuild.dev/zh/config/options/output.html#outputdisablecssmoduleextension) 来实现，比如：
+如果你想将其他 CSS 文件也当做 CSS Modules 模块进行处理，可以通过配置 [output.cssModules.auto](/config/options/output.html#cssmodulesauto) 来实现。
 
-```ts
-export default {
-  output: {
-    disableCssModuleExtension: true,
-  },
-};
-```
-
-设置后，以下两个文件都会被视为 CSS Modules：
-
-```ts
-import styles1 from './foo.module.css';
-import styles2 from './bar.css';
-```
-
-:::tip
-我们不推荐开启此配置项，因为开启 `disableCssModuleExtension` 后，CSS Modules 文件和普通 CSS 文件无法得到明确的区分，不利于长期维护。
-:::
-
-## 为指定的样式文件启用 CSS Modules
-
-在默认情况下，只有 `*.module.css` 结尾的文件才被视为 CSS Modules 模块。
-
-如果你想只为一些指定的样式文件启用 CSS Modules，可以通过配置 [output.cssModules](/config/options/output.html#outputcssmodule) 来实现，比如：
+比如：
 
 ```ts
 export default {
@@ -75,6 +52,13 @@ export default {
     },
   },
 };
+```
+
+设置后，以下两个文件都会被视为 CSS Modules：
+
+```ts
+import styles1 from './foo.module.css';
+import styles2 from './shared/bar.css';
 ```
 
 ## 自定义类名
@@ -133,35 +117,6 @@ declare module '*.module.styl' {
   export default classes;
 }
 declare module '*.module.stylus' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-```
-
-如果你开启了 `disableCssModuleExtension` 配置值，还需要添加以下类型：
-
-```ts title="src/env.d.ts"
-declare module '*.css' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-declare module '*.scss' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-declare module '*.sass' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-declare module '*.less' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-declare module '*.styl' {
-  const classes: { readonly [key: string]: string };
-  export default classes;
-}
-declare module '*.stylus' {
   const classes: { readonly [key: string]: string };
   export default classes;
 }
