+++ packages/document/docs/zh/shared/config/output/cssModules.md
@@ -15,11 +15,10 @@ type CssModules = {
 auto 配置项允许基于文件名自动启用 CSS 模块。
 
 - **类型：** `boolean | RegExp | ((resourcePath: string) => boolean)`
-- **默认值：** `undefined`
+- **默认值：** `true`
 
 类型说明：
 
-- `undefined`: 根据 [output.disableCssModuleExtension](https://rsbuild.dev/zh/config/options/output.html#outputdisablecssmoduleextension) 配置项决定为哪些样式文件启用 CSS 模块。
 - `true`: 为所有匹配 `/\.module\.\w+$/i.test(filename)` 正则表达式的文件启用 CSS 模块。
 - `false`: 禁用 CSS 模块。
 - `RegExp`: 为所有匹配 `/RegExp/i.test(filename)` 正则表达式的文件禁用 CSS 模块。
